requirements.txt
pyzmq
PyGObject

opencv-python
numpy